requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
requests>=2.31.0
scipy>=1.12.0
gdown>=4.7.3
python-dotenv>=1.0.0
